requirements.txt
anthropic==0.54.0
fastmcp==2.9.0
httpx==0.28.1
python-dotenv==1.1.0
playwright==1.45.0
playwright-stealth==1.0.6
beautifulsoup4==4.13.4
fake-useragent==2.2.0
pandas==2.2.1
typing_extensions==4.15.0
